Started cleaning up the code, comments, json4s library usage.
